+++ live/gateway.go
@@ -70,6 +70,7 @@ func NewChallengeRequestMsgFromBytes(b []byte) *ChallengeRequestMsg {
 
 // AuthenticationResponseMsg is an authentication response is sent by the gateway after a valid
 // authentication request is sent to the gateway.
+// https://databento.com/docs/api-reference-live/gateway-control-messages/authentication-response
 type AuthenticationResponseMsg struct {
 	Success   string // key: success
 	Error     string // key: error
@@ -93,14 +94,18 @@ func NewAuthenticationResponseMsgFromBytes(b []byte) *AuthenticationResponseMsg 
 
 // AuthenticationRequestMsg is an authentication request is sent to the gateway after a challenge response is received.
 // This is required to authenticate a user.
+// https://databento.com/docs/api-reference-live/client-control-messages/authentication-request
 type AuthenticationRequestMsg struct {
-	Auth              string        // key: auth
-	Dataset           string        // key: dataset
-	Encoding          dbn.Encoding  // key: encoding
-	Details           string        // key: details
-	TsOut             bool          // key: ts_out
-	Client            string        // key: client
-	HeartbeatInterval time.Duration // key: heartbeat_interval (in seconds); 0 means use server default
+	Auth               string                 // key: auth, required
+	Dataset            string                 // key: dataset, required
+	Client             string                 // key: client
+	Encoding           dbn.Encoding           // key: encoding
+	Compression        dbn.Compression        // key: compression
+	TsOut              bool                   // key: ts_out
+	PrettyPx           bool                   // key: pretty_px
+	PrettyTs           bool                   // key: pretty_ts
+	HeartbeatIntervalS uint32                 // key: heartbeat_interval_s (in seconds); 0 means use server default
+	SlowReaderBehavior dbn.SlowReaderBehavior // key: slow_reader_behavior, default "warn"
 }
 
 // NewAuthenticationRequestMsg parses a control message and returns a AuthenticationRequestMsg
@@ -112,17 +117,40 @@ func NewAuthenticationRequestMsg() AuthenticationRequestMsg {
 	}
 }
 
+const defaultGatewayEncoding = dbn.Encoding_Dbn
+const defaultGatewayCompression = dbn.Compress_None
+const defaultGatewaySlowReaderBehavior = dbn.SlowReaderBehavior_Warn
+
 // Encode converts AuthenticationRequestMsg to its line protocol representation.
 func (m *AuthenticationRequestMsg) Encode() []byte {
-	tsOutStr := "0"
-	if m.TsOut {
-		tsOutStr = "1"
+	b := fmt.Appendf(nil, "auth=%s|dataset=%s|client=%s",
+		m.Auth, m.Dataset, m.Client)
+	if m.Encoding != defaultGatewayEncoding {
+		if str := m.Encoding.String(); str != "" {
+			b = fmt.Appendf(b, "|encoding=%s", str)
+		}
 	}
-	b := fmt.Appendf(nil, "auth=%s|dataset=%s|encoding=%s|ts_out=%s|client=%s",
-		m.Auth, m.Dataset, m.Encoding.String(), tsOutStr, m.Client)
-	if m.HeartbeatInterval > 0 {
-		secs := int(m.HeartbeatInterval / time.Second)
-		b = fmt.Appendf(b, "|heartbeat_interval_s=%d", secs)
+	if m.Compression != defaultGatewayCompression {
+		if str := m.Compression.String(); str != "" {
+			b = fmt.Appendf(b, "|compressions=%s", str)
+		}
+	}
+	if m.TsOut {
+		b = fmt.Appendf(b, "|ts_out=1")
+	}
+	if m.PrettyPx {
+		b = fmt.Appendf(b, "|pretty_px=1")
+	}
+	if m.PrettyTs {
+		b = fmt.Appendf(b, "|pretty_ts=1")
+	}
+	if m.HeartbeatIntervalS != 0 {
+		b = fmt.Appendf(b, "|heartbeat_interval_s=%d", m.HeartbeatIntervalS)
+	}
+	if m.SlowReaderBehavior != defaultGatewaySlowReaderBehavior {
+		if str := m.SlowReaderBehavior.String(); str != "" {
+			b = fmt.Appendf(b, "|slow_reader_behavior=%s", str)
+		}
 	}
 	b = append(b, '\n')
 	return b